requirements.txt
opencv-python
transformers
torch
torchvision
Pillow
imutils
numpy<2.0
telepot
pyyaml
scikit-image
imagehash
pytest
Flask
timm
ultralytics
pandas
seaborn
